move update and delete api to the linksform  , add useeffect reset to the link card , seems fine . phewwwwwww
