+++ components/linksform.js
@@ -1,10 +1,15 @@
 import styles from "../styles/form.module.css";
 import LinkCard from "./linkcard";
 import { useStateValue } from "./context/state";
+import { useState } from "react";
 
-const LinksForm = ({ data, pagedataid }) => {
-  console.log(data);
+const endpoint =
+  process.env.NODE_ENV === "production" ? `` : "http://localhost:3000";
+
+const LinksForm = ({ pagedataid }) => {
+  // console.log(data);
   const [{ links }, dispatch] = useStateValue();
+  const [loading, setloading] = useState(false);
 
   const addNewLink = () => {
     // console.log(links.length);
@@ -31,6 +36,91 @@ const LinksForm = ({ data, pagedataid }) => {
     });
   };
 
+  const saveLinkData = async (linkdata) => {
+    console.log("save linkdata");
+    console.log(linkdata);
+    setloading(true);
+    // setshowAlert({
+    //   msg: "",
+    //   type: "",
+    // });
+
+    let operation = "insertpagelinks";
+    if (linkdata.hasOwnProperty("id")) {
+      operation = `updatepagelinks`;
+    }
+    console.log(operation);
+    try {
+      let res = await fetch(`${endpoint}/api/${operation}`, {
+        method: "POST",
+        body: JSON.stringify(linkdata),
+        headers: { "Content-Type": "application/json" },
+      }).then((res) => res.json());
+
+      // setshowAlert({
+      //   msg:
+      //     operation === "insertpagelinks"
+      //       ? "Added new page link "
+      //       : "Updated page link " + " successfully",
+      //   type: "success",
+      // });
+      console.log(res);
+      dispatch({ type: "updateLink", linkdata: res.updatedLinkData });
+      // reset(res.updatedLinkData[index]);
+    } catch (error) {
+      console.log(error);
+      // setshowAlert({
+      //   msg: operation + "failed" + error.message,
+      //   type: "danger",
+      // });
+    }
+    setloading(false);
+  };
+
+  const deleteLink = async (id) => {
+    console.log("delete link");
+    console.log(id);
+    setloading(true);
+    // setshowAlert({
+    //   msg: "",
+    //   type: "",
+    // });
+
+    let operation = "deletepagelink";
+    // if (linkdata.hasOwnProperty("id")) {
+    //   operation = `updatepagelinks`;
+    // }
+    console.log(operation);
+    try {
+      let res = await fetch(`${endpoint}/api/${operation}`, {
+        method: "POST",
+        body: JSON.stringify({ id: id }),
+        headers: { "Content-Type": "application/json" },
+      }).then((res) => res.json());
+
+      // setshowAlert({
+      //   msg:
+      //     operation === "insertpagelinks"
+      //       ? "Added new page link "
+      //       : "Updated page link " + " successfully",
+      //   type: "success",
+      // });
+      console.log(res);
+      // reset();
+      dispatch({ type: "deleteLink", id: id });
+      // console.warn("reSETTTTTTTTTTTTT");
+      // console.log(item);
+      // reset({});
+    } catch (error) {
+      console.log(error);
+      // setshowAlert({
+      //   msg: operation + "failed" + error.message,
+      //   type: "danger",
+      // });
+    }
+    setloading(false);
+  };
+
   return (
     <>
       <div className={styles.Wrapper}>
@@ -43,8 +133,6 @@ const LinksForm = ({ data, pagedataid }) => {
             type="button"
             className="btn btn-outline-primary"
             onClick={(e) => {
-              // console.log(e);
-              // console.log("click");
               addNewLink();
             }}
           >
@@ -55,10 +143,10 @@ const LinksForm = ({ data, pagedataid }) => {
               return (
                 <LinkCard
                   key={index}
-                  index={index}
                   item={item}
-                  // save={update}
-                  // loading={loading}
+                  deleteLink={deleteLink}
+                  updateLink={saveLinkData}
+                  loading={loading}
                 />
               );
             })}
